install: make the apt management a little more flexible
Allow the user to specify what file in /etc/apt/source.list.d/ to use. For
the local users we just always do Ceph, but we're about to use a bunch
of these functions for handling Samba separately.
(Does this mean we should pull these functions out of the task and into
a library of some kind?)

Signed-off-by: Greg Farnum <greg@inktank.com>