task: Restart task for testing daemon kill points
The ceph daemons support being killed at a specific code point
with a config option.  In some cases, we want to test a code point
that we can only set through the config file (the code point occurs
only on startup).  This task allows running a script or executable
and (when the script sends a command to the task) restarting it with
a temporary config that has the appropriate kill point set.  Once
the daemon asserts and gets restarted, the original config is used.

Right now this task follows the workunit task closely, but uses stdout/stdin
to specify when to restart a daemon.

Signed-off-by: Sam Lang <sam.lang@inktank.com>