task/daemon-helper: Add nostdin option
Some daemons (smbd) will try to read from stdin and check if its a
socket, using that for sending/receiving messages.  If /dev/null is
used for stdin, the daemon aborts.  This patch adds a 'nostdin' option
to the daemon-helper so that the daemon can be started without /dev/null
as stdin.

Signed-off-by: Sam Lang <sam.lang@inktank.com>