Don't specify ceph package version on yum command line.
It messes with obseleted packages. I think we did this in the first
place as an incorrect workaround. Since setting the priorities to
config file to check obseletes doing manually testing i found
things to install the correct version now without the version
being explicitly set in the command line so priorities is working
as expected.

Signed-off-by: Sandon Van Ness <sandon@redhat.com>